API reference

API reference

Single-limb arithmetic

DarkIntegers.addhilo — Function.

addhilo(x_hi::T, x_lo::T, y::T) where T <: Unsigned

Calculates hi * B + lo = x_hi * B + x_lo + y, where B == typemax(T) + 1. Returns the result as a pair (hi::T, lo::T). An overflow in hi (if any) is ignored.

source

DarkIntegers.mulhilo — Function.

mulhilo(x::T, y::T) where T <: Unsigned

Calculates hi * B + lo = x * y, where B == typemax(T) + 1. Returns the result as a pair (hi::T, lo::T).

source

DarkIntegers.divremhilo — Function.

divremhilo(x_hi::T, x_lo::T, y::T) where T <: Unsigned

Calculates divrem(x_hi * B + x_lo, y), where B == typemax(T) + 1. Returns a tuple (q::T, r::T, o::Bool), where q is the quotient (the part of it fitting into the bits of T), r is the remainder is o is the overflow flag.

source

DarkIntegers.divhilo — Function.

divhilo(x_hi::T, x_lo::T, y::T) where T <: Unsigned

Calculates div(x_hi * B + x_lo, y), where B == typemax(T) + 1. Returns a tuple (q::T, o::Bool), where q is the quotient (the part of it fitting into the bits of T), and o is the overflow flag.

source

DarkIntegers.remhilo — Function.

remhilo(x_hi::T, x_lo::T, y::T) where T <: Unsigned

Calculates rem(x_hi * B + x_lo, y), where B == typemax(T) + 1.

source

Single-limb modulo arithmetic

DarkIntegers.addmod — Function.

addmod(x::T, y::T, modulus::T) where T <: Unsigned

Returns mod(x + y, modulus) (even if x + y overflows T). Assumes x and y are in range 0 ... modulus.

source

DarkIntegers.submod — Function.

submod(x::T, y::T, modulus::T) where T <: Unsigned

Returns mod(x - y, modulus) (even if x - y underflows).

source

DarkIntegers.mulmod — Function.

mulmod(x::T, y::T, modulus::T) where T <: Unsigned

Returns mod(x * y, modulus) (even if x * y overflows T).

source

Multi-precision numbers

DarkIntegers.MPNumber — Type.

Multi-precision unsigned integer type, with N limbs of type T (which must be an unsigned integer type).

Supports +, -, *, divrem, div, rem, ^, <, <=, >, >=, zero, one and isodd.

MPNumber{N, T}(x::Integer) where {N, T <: Unsigned}

Creates an MPNumber object. If x does not fit into N limbs of type T, the excess bits will be ignored.

source

Residue ring elements

DarkIntegers.RRElem — Type.

Residue ring element, an unsigned integer with all operations performed modulo M.

Supports +, -, *, divrem, div, rem, ^, <, <=, >, >=, zero, one and isodd. Note that the division is a regular division, not multiplication by the inverse (which is not guaranteed to exist for any M).

RRElem{T, M}(x::Integer) where {T <: Unsigned, M}

Creates an RRElem object. M must have the type T.

source

Residue ring elements (Montgomery representation)

DarkIntegers.RRElemMontgomery — Type.

Residue ring element in Montgomery representation. Multiplication is much faster than for RRElem, addition and subtraction are the same, but division and conversion to regular integers is slower.

Supports +, -, *, divrem, div, rem, ^, <, <=, >, >=, zero, one and isodd. Note that the division is a regular division, not multiplication by the inverse (which is not guaranteed to exist for any M).

RRElemMontgomery{T, M}(x::Integer) where {T <: Unsigned, M}

Creates an RRElemMontgomery object. M must have the type T and be an odd number.

source

Cyclic polynomials

DarkIntegers.Polynomial — Type.

Polynomials modulo x^n-1 (cyclic) or x^n+1 (negacyclic). Supports any type that has arithmetic operators defined for it, including RRElem and RRElemMontgomery.

Polynomial(coeffs::AbstractArray{T, 1}, negacyclic::Bool) where T

Create a polynomial given the array of coefficients (the i-th coefficient corresponds to the (i-1)-th power).

source

DarkIntegers.shift_polynomial — Function.

shift_polynomial(p::Polynomial{T}, shift::Integer) where T

Multiply the polynomial by x^shift. shift can be negative.

source

Partial modification of residue ring elements and polynomials

DarkIntegers.change_representation — Function.

change_representation(
    new_repr::Union{RRElem, RRElemMontgomery},
    x::Union{RRElem{T, M}, RRElemMontgomery{T, M}}) where {T, M}

Change the representation of the given residue ring element to one of RRElem, RRElemMontgomery.

source

change_representation(new_repr, p::Polynomial{T}) where T <: AbstractRRElem

Apply change_representation() to every coefficient of the polynomial.

source

DarkIntegers.change_base_type — Function.

change_base_type(::Type{V}, x::RRElem{T, M}) where {T, M, V <: Unsigned}

Change the base type (T) of the given residue ring element to the type V. The modulus M must fit into V.

source

change_base_type(::Type{V}, p::Polynomial{T}) where {T <: AbstractRRElem, V <: Unsigned}

Apply change_base_type() to every coefficient of the polynomial.

source

DarkIntegers.change_modulus — Function.

change_modulus(new_modulus::Unsigned, x::RRElem{T, M}) where {T, M}

Change modulus to new_modulus (it will be converted to type T), keeping the value intact. If the new modulus is smaller than the current one, the value will be taken modulo new_modulus.

source

change_modulus(new_modulus::Unsigned, p::Polynomial{T}) where T <: AbstractRRElem

Apply change_modulus() to every coefficient of the polynomial.

source

DarkIntegers.change_length — Function.

change_length(new_length::Integer, p::Polynomial)

Change length of the polynomial, padding it with zeros. The new length must be greater or equal to the current length.

source

DarkIntegers.rescale — Function.

rescale(new_max::Unsigned, x::RRElem{T, M}, round_result::Bool)

Rescale x proportionally to the range [0, new_max) (where new_max <= M). Equivalent to floor(x * new_max / M) or round(...), depending on the value of round_result. If round_result is true, and the value if equal to new_max after rounding, it is set to 0.

source

rescale(new_max::Unsigned, p::Polynomial{T}, round_result::Bool) where T <: RRElem

Apply rescale() to every coefficient of the polynomial.

source

Miscellaneous

DarkIntegers.encompassing_type — Function.

encompassing_type(tp::Unsigned)

Returns the built-in type that covers all the range of tp. Works on UInt*, MPNumber, RRElem and RRElemMontgomery.

source