API reference

Utility functions

DarkIntegers.bitsizeof — Function

bitsizeof(::Type{T}) where T
bitsizeof(x)

Size, in bits, of the canonical binary representation of the given type T. Size, in bits, of object x if it is not a type.

DarkIntegers.log_bitsizeof — Function

log_bitsizeof(::Type{T}) where T
log_bitsizeof(x)

log2() of the value returned by bitsizeof for that argument.

DarkIntegers.num_bits — Function

num_bits(x)

Returns the number of bits in the representation of the absolute value of an integer.

DarkIntegers.encompassing_type — Function

encompassing_type(tp::Type{<:Unsigned})

Returns the built-in type that covers all the range of tp (not necessarily unsigned).

Single-limb arithmetic

DarkIntegers.addhilo — Function

addhilo(x_hi::T, x_lo::T, y::T) where T <: Unsigned

Calculates hi * B + lo = x_hi * B + x_lo + y, where B == typemax(T) + 1. Returns the result as a pair (hi::T, lo::T). An overflow in hi (if any) is ignored.

DarkIntegers.mulhilo — Function

mulhilo(x::T, y::T) where T <: Unsigned

Calculates hi * B + lo = x * y, where B == typemax(T) + 1. Returns the result as a pair (hi::T, lo::T).

DarkIntegers.divremhilo — Function

divremhilo(x_hi::T, x_lo::T, y::T) where T <: Unsigned

Calculates divrem(x_hi * B + x_lo, y), where B == typemax(T) + 1. Returns a tuple (q::T, r::T, o::Bool), where q is the quotient (the part of it fitting into the bits of T), r is the remainder is o is the overflow flag.

DarkIntegers.divhilo — Function

divhilo(x_hi::T, x_lo::T, y::T) where T <: Unsigned

Calculates div(x_hi * B + x_lo, y), where B == typemax(T) + 1. Returns a tuple (q::T, o::Bool), where q is the quotient (the part of it fitting into the bits of T), and o is the overflow flag.

DarkIntegers.remhilo — Function

remhilo(x_hi::T, x_lo::T, y::T) where T <: Unsigned

Calculates rem(x_hi * B + x_lo, y), where B == typemax(T) + 1.

Single-limb modulo arithmetic

DarkIntegers.addmod — Function

addmod(x::T, y::T, modulus::T) where T <: Unsigned

Returns mod(x + y, modulus) (even if x + y overflows T). Assumes x and y are in range 0 ... modulus.

DarkIntegers.submod — Function

submod(x::T, y::T, modulus::T) where T <: Unsigned

Returns mod(x - y, modulus) (even if x - y underflows).

DarkIntegers.mulmod — Function

mulmod(x::T, y::T, modulus::T) where T <: Unsigned

Returns mod(x * y, modulus) (even if x * y overflows T).

Multi-precision numbers

Missing docstring.

Missing docstring for MLUint. Check Documenter's build log for details.

Residue ring elements

DarkIntegers.ModUInt — Type

Residue ring element, an unsigned integer with all operations performed modulo M.

Supports +, -, *, divrem, div, rem, ^, <, <=, >, >=, zero, one and isodd. Note that the division is a regular division, not multiplication by the inverse (which is not guaranteed to exist for any M).

ModUInt{T, M}(x::Integer) where {T <: Unsigned, M}

Creates an ModUInt object. M must have the type T.

Residue ring elements (Montgomery representation)

DarkIntegers.MgModUInt — Type

Residue ring element in Montgomery representation. Multiplication is much faster than for ModUInt, addition and subtraction are the same, but division and conversion to regular integers is slower.

Supports +, -, *, divrem, div, rem, ^, <, <=, >, >=, zero, one and isodd. Note that the division is a regular division, not multiplication by the inverse (which is not guaranteed to exist for any M).

MgModUInt{T, M}(x::Integer) where {T <: Unsigned, M}

Creates an MgModUInt object. M must have the type T and be an odd number.

Cyclic polynomials

DarkIntegers.Polynomial — Type

Fixed-size polynomials (with the degree limited by N-1). Currently supports moduli x^n-1 (cyclic_modulus) or x^n+1 (negacyclic_modulus). Supports any type that has arithmetic operators defined for it, including ModUInt and MgModUInt.

Polynomial(coeffs::AbstractArray{T, 1}, modulus::AbstractCyclicModulus) where T

Create a polynomial given the array of coefficients (the i-th coefficient corresponds to the (i-1)-th power).

DarkIntegers.mul_by_monomial — Function

mul_by_monomial(p::Polynomial, power::Integer)

Multiply the polynomial by x^power. If power lies outside [0, 2 * N) where N-1 is the maximum degree of the polynomial, a modulo 2 * N will be taken.

Partial modification of residue ring elements and polynomials

Missing docstring.

Missing docstring for change_representation. Check Documenter's build log for details.

Missing docstring.

Missing docstring for change_base_type. Check Documenter's build log for details.

Missing docstring.

Missing docstring for change_modulus. Check Documenter's build log for details.

Missing docstring.

Missing docstring for change_length. Check Documenter's build log for details.

Missing docstring.

Missing docstring for rescale. Check Documenter's build log for details.